requirements.txt
aniso8601==9.0.1
click==8.1.3
colorama==0.4.5
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.1
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
Flask==2.1.2
greenlet==1.1.2
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysqlclient==2.1.0
pip==21.1.2
PyJWT==2.4.0
python-dotenv==0.20.0
pytz==2022.1
setuptools==57.0.0
six==1.16.0
SQLAlchemy==1.4.37
Werkzeug==2.1.2
wheel==0.36.2
zipp==3.8.0
